refactor: jpa 적용
