+++ src/main/java/org/sopt/domain/member/dto/request/MemberCreateDto.java
@@ -1,0 +1,33 @@
+package org.sopt.domain.member.dto.request;
+
+import java.time.LocalDate;
+
+import org.sopt.domain.member.entity.Gender;
+import org.sopt.domain.member.entity.Member;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record MemberCreateDto(
+	@NotBlank
+	String name,
+	@Email
+	String email,
+	@NotNull
+	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+	LocalDate birthday,
+	@NotNull
+	Gender gender
+) {
+	public Member toEntity() {
+		return Member.builder()
+				.name(name)
+				.email(email)
+				.birthday(birthday)
+				.gender(gender)
+				.build();
+	}
+}
