+++ src/main/java/org/sopt/domain/member/entity/Gender.java
@@ -1,0 +1,13 @@
+package org.sopt.domain.member.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Gender {
+	MALE("남성"),
+	FEMALE("여성");
+
+	private final String value;
+}
