+++ src/main/java/org/sopt/domain/member/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package org.sopt.domain.member.repository;
+
+import org.sopt.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+	boolean existsByEmail(String email);
+}
