+++ src/main/java/org/sopt/domain/member/service/MemberService.java
@@ -1,9 +1,9 @@
-package org.sopt.member.service;
+package org.sopt.domain.member.service;
 
 import java.util.List;
 
-import org.sopt.member.dto.request.MemberCreateDto;
-import org.sopt.member.dto.response.MemberInfoDto;
+import org.sopt.domain.member.dto.request.MemberCreateDto;
+import org.sopt.domain.member.dto.response.MemberInfoDto;
 
 public interface MemberService {
