--- src/main/java/org/sopt/member/domain/Gender.java
@@ -1,25 +1,0 @@
-package org.sopt.member.domain;
-
-import java.util.Arrays;
-
-public enum Gender {
-	MALE("남성"),
-	FEMALE("여성");
-
-	private final String value;
-
-	Gender(String value) {
-		this.value = value;
-	}
-
-	public String getValue() {
-		return value;
-	}
-
-	public static Gender of(String value) {
-		return Arrays.stream(Gender.values())
-				.filter(g -> g.value.equals(value))
-				.findFirst()
-				.orElseThrow(NullPointerException::new);
-	}
-}
