--- src/main/java/org/sopt/member/repository/MemberRepository.java
@@ -1,20 +1,0 @@
-package org.sopt.member.repository;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.sopt.member.domain.Member;
-
-public interface MemberRepository {
-	void init();
-
-	Member save(Member member);
-
-	Optional<Member> findById(Long id);
-
-	List<Member> findAll();
-
-	void deleteById(Long id);
-
-	boolean existsByEmail(String email);
-}
