--- src/main/java/org/sopt/member/repository/util/DataInitializer.java
@@ -1,20 +1,0 @@
-package org.sopt.member.repository.util;
-
-import org.sopt.member.repository.MemberRepository;
-import org.springframework.stereotype.Component;
-
-import jakarta.annotation.PostConstruct;
-
-@Component
-public class DataInitializer {
-	private final MemberRepository memberRepository;
-
-	public DataInitializer(MemberRepository memberRepository) {
-		this.memberRepository = memberRepository;
-	}
-
-	@PostConstruct
-	public void init() {
-		memberRepository.init();
-	}
-}
