--- src/main/java/org/sopt/member/repository/util/ExecutorManager.java
@@ -1,18 +1,0 @@
-package org.sopt.member.repository.util;
-
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-public class ExecutorManager {
-	private static final ExecutorService executor = Executors.newSingleThreadExecutor();
-
-	private ExecutorManager() {}
-
-	public static ExecutorService getExecutor() {
-		return executor;
-	}
-
-	public static void shutdown() {
-		executor.shutdown();
-	}
-}
